[CHORE/#1547] MissionDetail 파일 위치 및 네이밍 수정
