+++ feature/soptamp/src/main/java/org/sopt/official/stamp/feature/mission/detail/type/MissionDetailModeType.kt
@@ -1,0 +1,7 @@
+package org.sopt.official.stamp.feature.mission.detail.type
+
+enum class MissionDetailModeType {
+    WRITE,
+    READ_ONLY,
+    EDIT,
+}
